Keypad Interface Game Design Using 8051
Games are mostly fascinating to play and there exists much engineering thoughts behind every logic in the game. The present day world deals with the games which are very precise in results in terms of speed, timing and placements. With this project we come with a new game where a person’s timing and observational skills are tested and awarded with a score. In this project we use 8051 micro-controller which is interfaced with 7 segment display, a 7 x 16 LCD, and a Hex Keypad. The micro-controller generates some random number on the 7 segment. The conditions of the game is that we have to press the same number that is displayed in a stipulated amount of time say 1 second and send the same  number using external keypad to 8051 before it displays the next random number. So if we fail to enter the number in the given time the game ends by displaying your result on the LCD. The delay time will be initiated with certain clock frequency and the frequency goes on increasing d