Library Management System
The main aim of this project is to implement an application which deals with maintaining Library activities like Book issues, book returns and Administration activities. Initially all the information about books like titles and volumes will be entered and maintained. This system will reduce manual work for maintaining records in files.
This Application maintains the history of usage of the books. This system provides the search facility to search the status of the books. Regular transactions which include book issue, returns etc. and exceptional transactions that are related to loss of books, damage of books etc. also will have to be handled by the system.
Our project is only a humble venture to satisfy the needs in a library. Several user friendly coding have also adopted. This package shall prove to be a powerful package in satisfying all the requirements of the organization.