Count number of 1's and 0's in a given sequence
A binary sequence contains either 1’s or 0’s. It is necessary to count number of one’s and zero’s in the given sequence. As while transmitting we can have a count of how many one’s and zero’s we are transmitting. If we have counter even at receiver we can count the number of one’s and zero’s that are received. So that we can check whether the count of the given sequence while transmitting equal to the count at the receiver. In order to develop a counter for counting number of one’s and zero’s in the given sequence we need to have two ripple counters (one to count the occurrence of one’s and other for zero’s), a shift register which stores the given binary sequence and shifts the bits as soon as the input bit stream is loaded into register. A BCD to 7-segment decoder is also used to display the count of one’s and zero’s in 7- segment display.