Bidirectional Visitor Counter using 8051
The project BIDIRECTIONAL VISITOR COUNTING SYSTEM helps us in maintaining details about the number of people entering as well as leaving a particular place such as a seminar hall, conference hall or other public places. The main concept behind this is to monitor the number of people inside the place in-order to avoid congestion.
In this project we use 3 switches, one for maintaining the count of people entering and another is used to maintain the count of people leaving that particular place. The third switch is used to check the number of persons inside.
The value of the register connected to the switch (entering) and the register maintaining the number of people inside are incremented when a person enters and the switch connected(leaving) is incremented and the register maintaining the number of people inside is decremented if a person leaves the place.
The number of people entering and leaving the place will be displayed on the LCD and it can also be used to dis